refactor: separate logic in multiple modules
