+++ simplemma/__init__.py
@@ -10,3 +10,5 @@ __version__ = "0.9.0"
 from .langdetect import in_target_language, lang_detector
 from .simplemma import lemmatize, lemma_iterator, text_lemmatizer, is_known
 from .tokenizer import simple_tokenizer
+from .dictionaries import DictionaryCache
+from .dictionary_pickler import *
