+++ simplemma/utils.py
@@ -1,0 +1,33 @@
+from functools import lru_cache
+
+
+@lru_cache(maxsize=65536)
+def levenshtein_dist(str1: str, str2: str) -> int:
+    # inspired by this noticeably faster code:
+    # https://gist.github.com/p-hash/9e0f9904ce7947c133308fbe48fe032b
+    if str1 == str2:
+        return 0
+    if len(str1) > len(str2):
+        str1, str2 = str2, str1
+    r1 = list(range(len(str2) + 1))
+    r2 = [0] * len(r1)
+    for i, c1 in enumerate(str1):
+        r2[0] = i + 1
+        for j, c2 in enumerate(str2):
+            if c1 == c2:
+                r2[j + 1] = r1[j]
+            else:
+                a1, a2, a3 = r2[j], r1[j], r1[j + 1]
+                if a1 > a2:
+                    if a2 > a3:
+                        r2[j + 1] = 1 + a3
+                    else:
+                        r2[j + 1] = 1 + a2
+                else:
+                    if a1 > a3:
+                        r2[j + 1] = 1 + a3
+                    else:
+                        r2[j + 1] = 1 + a1
+        aux = r1
+        r1, r2 = r2, aux
+    return r1[-1]
